00.01.02.01

+++ src/MKTDLL/dllDefines.h
@@ -23,5 +23,5 @@ typedef struct functionInput {
 } funi;
 
 typedef struct cleanUpInput {
-    int * errorCode;
+    long int * errorCode;
 } cleanUpi;
